Split stacking code out in separate module
