+++ karabo_data/__init__.py
@@ -37,8 +37,9 @@ __version__ = "0.5.0"
 
 
 from .reader import *
+from .stacking import *
 from .export import *
 from .utils import *
 
 
-__all__ = export.__all__ + reader.__all__ + utils.__all__
+__all__ = export.__all__ + reader.__all__ + utils.__all__ + stacking.__all__
