+++ karabo_data/read_machinery.py
@@ -205,86 +205,3 @@ def find_proposal(propno):
         return d
 
     raise Exception("Couldn't find proposal dir for {!r}".format(propno))
-
-
-class VirtualStack:
-    """'Virtual' array for stacking detector data with modules on axis -3
-
-    Access is limited to either a single module at a time or all modules
-    together, but this is enough to assemble detector images.
-    """
-    def __init__(self, data, nmodules, mod_shape, dtype, fillvalue,
-                 stack_axis=-3):
-        self._nmodules = nmodules
-        self._data = data  # {modno: array}
-        self.dtype = dtype
-        self._fillvalue = fillvalue
-        self._mod_shape = mod_shape
-        self.ndim = len(mod_shape) + 1
-        self._stack_axis = stack_axis
-        if self._stack_axis < 0:
-            self._stack_axis += self.ndim
-        sax = self._stack_axis
-        self.shape = mod_shape[:sax] + (nmodules,) + mod_shape[sax:]
-
-    def __repr__(self):
-        return "<VirtualStack (shape={}, {}/{} modules, dtype={})>".format(
-            self.shape, len(self._data), self._nmodules, self.dtype,
-        )
-
-    # Multidimensional slicing
-    def __getitem__(self, item):
-        if not isinstance(item, tuple):
-            item = (item,)
-
-        missing_dims = self.ndim - len(item)
-        if Ellipsis in item:
-            ix = item.index(Ellipsis)
-            missing_dims += 1
-            item = item[:ix] + (slice(None, None),) * missing_dims + item[ix+1:]
-        else:
-            item = item + (slice(None, None),) * missing_dims
-
-        modno = item[self._stack_axis]
-        mod_slices = item[:self._stack_axis] + item[self._stack_axis+1:]
-
-        if isinstance(modno, int):
-            if modno < 0:
-                modno += self._nmodules
-            return self._get_single_mod(modno, mod_slices)
-        elif modno == slice(None, None):
-            return self._get_all_mods(mod_slices)
-        else:
-            raise Exception(
-                "VirtualStack can only slice a single module or all modules"
-            )
-
-    def _get_single_mod(self, modno, mod_slices):
-        try:
-            mod_data = self._data[modno]
-        except KeyError:
-            if modno > self._nmodules:
-                raise IndexError(modno)
-            mod_data = np.full(self._mod_shape, self._fillvalue, self.dtype)
-
-        # Now slice the module data as requested
-        return mod_data[mod_slices]
-
-    def _get_all_mods(self, mod_slices):
-        new_data = {modno: self._get_single_mod(modno, mod_slices)
-                    for modno in self._data}
-        new_mod_shape = list(new_data.values())[0].shape
-        return VirtualStack(new_data, self._nmodules, new_mod_shape, self.dtype,
-                            self._fillvalue)
-
-    def asarray(self):
-        """Copy this data into a real numpy array
-
-        Don't do this until necessary - the point of using VirtualStack is to
-        avoid copying the data unnecessarily.
-        """
-        start_shape = (self._nmodules,) + self._mod_shape
-        arr = np.full(start_shape, self._fillvalue, dtype=self.dtype)
-        for modno, data in self._data.items():
-            arr[modno] = data
-        return np.moveaxis(arr, 0, self._stack_axis)
