+++ karabo_data/reader.py
@@ -23,7 +23,6 @@ import pandas as pd
 import re
 import sys
 import tempfile
-from warnings import warn
 import xarray
 
 from .exceptions import SourceNameError, PropertyNameError, TrainIDError
@@ -36,7 +35,6 @@ from .read_machinery import (
     union_selections,
     contiguous_regions,
     find_proposal,
-    VirtualStack,
 )
 
 __all__ = [
@@ -45,8 +43,6 @@ __all__ = [
     'open_run',
     'FileAccess',
     'DataCollection',
-    'stack_data',
-    'stack_detector_data',
     'by_id',
     'by_index',
     'SourceNameError',
@@ -1212,122 +1208,3 @@ def open_run(proposal, run, data='raw'):
         run = 'r' + run.rjust(4, '0')
 
     return RunDirectory(osp.join(prop_dir, data, run))
-
-
-def stack_data(train, data, axis=-3, xcept=()):
-    """Stack data from devices in a train.
-
-    For detector data, use stack_detector_data instead: it can handle missing
-    modules, which this function cannot.
-
-    The returned array will have an extra dimension. The data will be ordered
-    according to any groups of digits in the source name, interpreted as
-    integers. Other characters do not affect sorting. So:
-
-        "B_7_0" < "A_12_0" < "A_12_1"
-
-    Parameters
-    ----------
-    train: dict
-        Train data.
-    data: str
-        The path to the device parameter of the data you want to stack.
-    axis: int, optional
-        Array axis on which you wish to stack.
-    xcept: list
-        List of devices to ignore (useful if you have reccored slow data with
-        detector data in the same run).
-
-    Returns
-    -------
-    combined: numpy.array
-        Stacked data for requested data path.
-    """
-    devices = [dev for dev in train.keys() if dev not in xcept]
-
-    if not devices:
-        raise ValueError("No data after filtering by 'xcept' argument.")
-
-    dtypes, shapes = set(), set()
-    ordered_arrays = []
-    for device in sorted(devices, key=lambda d: list(map(int, re.findall(r'\d+', d)))):
-        array = train[device][data]
-        dtypes.add(array.dtype)
-        ordered_arrays.append(array)
-
-    if len(dtypes) > 1:
-        raise ValueError("Arrays have mismatched dtypes: {}".format(dtypes))
-
-    return np.stack(ordered_arrays, axis=axis)
-
-
-def stack_detector_data(train, data, axis=-3, modules=16, fillvalue=np.nan,
-                        virtual=False):
-    """Stack data from detector modules in a train.
-
-    Parameters
-    ----------
-    train: dict
-        Train data.
-    data: str
-        The path to the device parameter of the data you want to stack, e.g. 'image.data'.
-    axis: int
-        Array axis on which you wish to stack (default is -3).
-    modules: int
-        Number of modules composing a detector (default is 16).
-    fillvalue: number
-        Value to use in place of data for missing modules. The default is nan
-        (not a number) for floating-point data, and 0 for integers.
-    virtual: bool
-        If True, avoid copying the data by making a 'virtual' stack.
-        This is limited compared to a real array, but sufficient for passing
-        to the geometry machinery to assemble images.
-
-    Returns
-    -------
-    combined: numpy.array
-        Stacked data for requested data path.
-    """
-
-    if not train:
-        raise ValueError("No data")
-
-    dtypes, shapes, empty_mods = set(), set(), set()
-    modno_arrays = {}
-    for device in train:
-        det_mod_match = re.search(r'/DET/(\d+)CH', device)
-        if not det_mod_match:
-            raise ValueError("Non-detector source: {}".format(device))
-        modno = int(det_mod_match.group(1))
-
-        try:
-            array = train[device][data]
-        except KeyError:
-            continue
-        dtypes.add(array.dtype)
-        shapes.add(array.shape)
-        modno_arrays[modno] = array
-
-    if len(dtypes) > 1:
-        raise ValueError("Arrays have mismatched dtypes: {}".format(dtypes))
-    if len(shapes) > 1:
-        s1, s2, *_ = sorted(shapes)
-        if len(shapes) > 2 or (s1[0] != 0) or (s1[1:] != s2[1:]):
-            raise ValueError("Arrays have mismatched shapes: {}".format(shapes))
-        empty_mods = {n for n, a in modno_arrays.items() if a.shape == s1}
-        for modno in empty_mods:
-            del modno_arrays[modno]
-        shapes.remove(s1)
-    if max(modno_arrays) >= modules:
-        raise IndexError("Module {} is out of range for a detector with {} modules"
-                         .format(max(modno_arrays), modules))
-
-    dtype = dtypes.pop()
-    shape = shapes.pop()
-    stack = VirtualStack(
-        modno_arrays, modules, shape, dtype, fillvalue, stack_axis=axis
-    )
-    if virtual:
-        return stack
-
-    return stack.asarray()
